Moved message structs into a Message folder
Each message now also has an own script
