+++ Assets/Scripts/Networking/Messages/MessageNetworkBehaviourInitialized.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e9322b3c4c140b46a1eaa692402c4d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
