+++ Assets/Scripts/Networking/Messages/MessageServerObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db54b813ccfe12b4d85a769fcaeac6ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
