+++ Assets/Scripts/Networking/Messages/MessageServerObjectList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c6001657ad55a14bab1d35823ccbe25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
